feat: Split reference and smart picker functions from components
