+++ src/components/NcRichText/index.js
@@ -22,26 +22,26 @@
 
 import NcRichText from './NcRichText.vue'
 
-import { registerWidget, renderWidget, isWidgetRegistered } from './widgets.js'
+import { registerWidget, renderWidget, isWidgetRegistered } from './../../functions/reference/widgets.js'
 import NcReferenceList from './NcReferenceList.vue'
 import NcReferenceWidget from './NcReferenceWidget.vue'
 import NcReferencePicker from './NcReferencePicker/NcReferencePicker.vue'
 import NcReferencePickerModal from './NcReferencePicker/NcReferencePickerModal.vue'
 import NcSearch from './NcReferencePicker/NcSearch.vue'
-import { getLinkWithPicker } from './NcReferencePicker/referencePickerModal.js'
+import { getLinkWithPicker } from './../../functions/reference/referencePickerModal.js'
 import {
 	getProvider,
 	getProviders,
 	sortProviders,
 	searchProvider,
 	anyLinkProviderId,
-} from './NcReferencePicker/providerHelper.js'
+} from '../../functions/reference/providerHelper.js'
 import {
 	registerCustomPickerElement,
 	renderCustomPickerElement,
 	isCustomPickerElementRegistered,
 	NcCustomPickerRenderResult,
-} from './NcReferencePicker/customPickerElements.js'
+} from '../../functions/reference/customPickerElements.js'
 
 export default NcRichText
