+++ src/functions/index.js
@@ -1,3 +1,4 @@
 export * from './a11y/index.ts'
 export * from './emoji/index.ts'
+export * from './reference/index.js'
 export { default as usernameToColor } from './usernameToColor/index.js'
