+++ src/functions/registerReference/index.js
@@ -1,0 +1,14 @@
+export {
+	registerWidget,
+	renderWidget,
+	destroyWidget,
+	isWidgetRegistered,
+} from './../reference/widgets.js'
+export {
+	NcCustomPickerRenderResult,
+	registerCustomPickerElement,
+	renderCustomPickerElement,
+	destroyCustomPickerElement,
+	isCustomPickerElementRegistered,
+	getCustomPickerElementSize,
+} from './../reference/customPickerElements.js'
